+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,9 +1,9 @@
-## 3.0.0 (Released 2024-03-xx)
+## 3.0.0 (Released 2024-xx-xx)
 
 - Updated to .NET 8.
 - Added support for reading FLAC files.
 - Four strings are displayed for the bass tuning when "Show advanced features" is disabled.
-- When adding a bass arrangement, the tuning for the nonexistent strings is set to match the fourth string.
+- When adding a bass arrangement, the tuning for the nonexistent strings is set to match the tuning offset of the fourth string.
 - Added validation for low bass tuning issues.
 - Added support for Wwise 2023.
 - Fixed the loading of the Wwise template on Linux.